Extract url test methods and classes to separate test module
